Updated backtest approach
