+++ src/main/java/com/example/tradeservice/backtest/series/DoubleSeries.java
@@ -1,4 +1,4 @@
-package com.example.tradeservice.strategy.series;
+package com.example.tradeservice.backtest.series;
 
 import java.util.List;
 import java.util.function.Function;
