+++ src/main/java/com/example/tradeservice/backtest/series/TimeSeries.java
@@ -1,4 +1,4 @@
-package com.example.tradeservice.strategy.series;
+package com.example.tradeservice.backtest.series;
 
 import java.time.Instant;
 import java.util.*;
