+++ src/main/java/com/example/tradeservice/strategy/dataclient/StockDataClient.java
@@ -5,7 +5,5 @@ import com.example.tradeservice.model.enums.TimeFrame;
 
 public interface StockDataClient {
 
-    void initializeCsvForDay(String symbol, String date);
-
     TwelveCandleBar quoteWithInterval(String symbol, TimeFrame timeFrame, String date);
 }
