+++ src/main/java/com/example/tradeservice/strategy/dataclient/TwelveDataClient.java
@@ -49,14 +49,10 @@ public class TwelveDataClient implements StockDataClient {
                         + "&symbol=" + symbol.toUpperCase()
                         + "&interval=" + TimeFrame.ONE_DAY.getTwelveFormat()
                         + "&format=CSV"
+                        + "&outputsize=5000"
                 )
                 .retrieve()
                 .body(String.class);
-    }
-
-    @Override
-    public void initializeCsvForDay(String symbol, String date) {
-        //do nothing
     }
 
     @Override
